orders/requirements.txt
Django==3.2.21
djangorestframework==3.14.0
django_rest_passwordreset==1.3.0
drf_spectacular==0.26.5
django-environ==0.11.2
celery==5.3.4
pytest==7.4.2
pytest-django==4.5.2
psycopg2-binary==2.9.7
requests==2.31.0
PyYAML==6.0.1
redis==5.0.0
django-baton==2.8.0
django-silk==5.0.4
